backend/requirements.txt
fastapi
uvicorn
openai
twilio
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok
pydub
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain
langgraph
langchain-openai
langchain_tavily
langgraph-checkpoint-postgres